merge: #113 컴포넌트 합침
